+++ NEWS.md
@@ -1,1 +1,7 @@
+# shinyD3 0.99.01
+
+* Transform package for multiple example
+
 # shinyD3 0.99.0 
+
+* shiny app with first example